requirements.txt
google-genai
openai
anthropic
python-dotenv
